Alerts 3.3

+++ lib/models/alert_aux_model.dart
@@ -7,8 +7,8 @@ class AlertAuxModel {
     required this.estado,
   });
 
-  int latitud;
-  int longitud;
+  double latitud;
+  double longitud;
   int tipoIncidente;
   String estado;
